Added auth middleware and removed duplicate wallet code
